Crate ruroonga_command

ruroonga_command provides extensible Groonga command query builder and generator. It reduces runtime errors about Groonga queries.

Modules

cache_limit

column_copy

column_create

column_list

column_remove

column_rename

command
Type variant definitions for Groonga commands.

command_line
An entity type for commandable trait.

command_query
An entity type for queryable trait.

commandable
commandable provides a shortcut to create command line Groonga query.

database_unmap

delete

dsl
dsl provides its XXXCommand returning dsl syntax like this:

dump

extendable
extendable provides appending user-defined key-value pair query functionality trait.

extension
This module provides fully user-defined query creation mechanism. This module is only provided for funture extensibility. You should use this module at your own risk.

io_flush

load

log_level

log_put

log_reopen

normalize

normalizer_list

object_exist

object_inspect

object_remove

plugin_register

plugin_unregister

queryable
queryable provides a shortcut to create URI style Groonga query.

request_cancel
Using request cancel mechanism. If you use this cancel feature, you should use request_cancellable trait to add request_id into each request.

request_cancellable
request_cancellable provides appending request_id functionality trait.

request_timeoutable
request_timeoutable provides appending timeout functionality trait.

schema

select
SelectCommand type creates select query.

selectable
Provides manupilation way to construct and concatinate drilldown with select query. Provides drilldown related procedures. This module will be used with SelectCommand type by + operation to concatinate select and drilldown queries.

shutdown

status

table_create

table_list

table_remove

table_rename

thread_limit

tokenize

tokenizer_list

truncate

types
Provides types for Query builders.