Struct ruroonga_command::command_line::CommandLine

pub struct CommandLine {
    // some fields omitted
}

Methods

impl CommandLine

fn new(command: Command, query: Query) -> CommandLine

fn get_command(&mut self) -> Command

Get command enum constant to construct command line query.

fn get_argument(&mut self) -> Query

Get vectorize ("key", "value") pairs to construct command line query.

fn set_argument(&mut self, arguments: Query)

Set vectorize ("key", "value") pairs to construct command line query.

fn make_command(&mut self) -> String

Create command line style command query.

vec![("key","value")] interprets to "--key value". And two or more value pair are concatinate with (a space).

fn encode(&mut self) -> String

Create Groonga command line style command.