Struct ruroonga_command::command_query::CommandQuery

pub struct CommandQuery { /* fields omitted */ }

Methods

impl CommandQuery

fn new(command: Command, query: Query) -> CommandQuery

fn get_command(&mut self) -> Cow<Command>

Get command enum constant to construct url encoded query.

fn get_argument(&mut self) -> Cow<Query>

Get vectorize ("key", "value") pairs to construct url encoded query.

fn set_argument(&mut self, arguments: Query)

Set vectorize ("key", "value") pairs to construct url encoded query.

fn make_query(&mut self) -> String

Create url encoded command query.

vec![("key","value")] interprets to "key=value". And two or more value pair are concatinate with &.

fn encode(&mut self) -> String

Create Groonga HTTP server query URL.