Module ruroonga_command::selectable

Provides manupilation way to construct and concatinate drilldown with select query. Provides drilldown related procedures. This module will be used with SelectCommand type by + operation to concatinate select and drilldown queries.

Also this module support pseudo_table and staged_columns manupilation.

Modules

drilldown

drilldown_builder

drilldown_type

drilldownable

fragmentable
fragmentable module provides making query fragment operation. It is useful to make composable query elements. It makes select query to be composable against drilldown related query and staged_columns query.

labeled_drilldown

labeled_drilldown_builder

labeled_drilldown_sequencable

labeled_drilldown_sequence_builder

labeled_drilldownable

pseudo_table
pseudo_table is a trait to be used for tablable state which is appeared in "drilldown in drilldown".

staged_columnable

staged_columns
staged_columns provides making query which supports window function syntax.

staged_columns_builder